组织通讯录选取成员
Former-commit-id: e2cc26b88fac75dd6d2d3494a1ee5752272362de

+++ wfuikit/WFChatUIKit/CommonVC/WFCUProfileTableViewController.m
@@ -675,7 +675,7 @@
             //click momentCell
         }
         
-        int index = indexPath.row;
+        int index = (int)indexPath.row;
         if(self.momentCell)
             index--;
